+++ packages/renderer/src/views/ArcTreeView.vue
@@ -11,7 +11,7 @@ import ConfirmationDialog from '../dialogs/ConfirmationDialog.vue';
 import StringDialog from '../dialogs/StringDialog.vue';
 import AddProtocolDialog from '../dialogs/AddProtocolDialog.vue';
 import ProgressDialog from '../dialogs/ProgressDialog.vue';
-import GitDialog from '../dialogs/GitDialog.vue';
+import { useDownloadLFSFiles } from '../composables/useDownloadLFSFiles';
 import { useQuasar } from 'quasar'
 import {ArcStudy, ArcAssay, ArcRun, ArcWorkflow, DataMap as ArcDatamap, ArcInvestigation, ARC} from '@nfdi4plants/arctrl';
 import { setAssayIdentifier, setRunIdentifier, setStudyIdentifier, setWorkflowIdentifier } from "@nfdi4plants/arctrl/Core/IdentifierSetters";
@@ -29,7 +29,7 @@ const $q = useQuasar();
 
 type LFSJsonFileMap = Record<string, LFSJsonFile>
 
-interface ArcTreeViewNode {
+export interface ArcTreeViewNode {
     header: string;
     type: string;
     id: string;
@@ -77,7 +77,7 @@ function relPathToDatamapParentInfo(s: string): SCS.DatamapParentInfo {
   let parentIdentifier: string | null = null;
   if (parts.length === 3) {
       const [pt, pi, datamapFileName] = parts;
-      if (datamapFileName !== ArcDatamap_FileName) 
+      if (datamapFileName !== ArcDatamap_FileName)
         throw new Error("Invalid datamap file name: " + datamapFileName);
       parentType = pt;
       parentIdentifier = pi;
@@ -137,14 +137,6 @@ function datamapParentInfoToRelPath(info: SCS.DatamapParentInfo): string {
 function getDatamapParentByPath(arc: ARC, path: string): ArcAssay | ArcStudy | ArcRun | ArcWorkflow | null {
   const info = relPathToDatamapParentInfo(path);
   return SCS.getDatamapParentByInfo(arc, info.ParentId, info.Parent);
-}
-
-function recUpdateNodes(fn: (n: ArcTreeViewNode)=>void, nodes: ArcTreeViewNode[]) {
-  for(const n of nodes) {
-    fn(n);
-    if(n.children)
-      recUpdateNodes(fn,n.children);
-  }
 }
 
 let uniqueLabelCounter = 0;
@@ -448,19 +440,27 @@ const triggerNode = (e: any, node: ArcTreeViewNode) => {
     case formatNodeEditString(Markdown):
       props.selection = node.id;
       AppProperties.active_markdown = node.id;
-      return AppProperties.state=AppProperties.STATES.EDIT_MARKDOWN;
+      AppProperties.state = AppProperties.STATES.EDIT_MARKDOWN;
+      break;
     case formatNodeEditString(Image):
       props.selection = node.id;
       AppProperties.active_image = node.id;
-      return AppProperties.state=AppProperties.STATES.EDIT_IMAGE;
+      AppProperties.state = AppProperties.STATES.EDIT_IMAGE;
+      break;
     case _NODE_TYPE_FALLBACK_:
       props.selection = node.id;
       AppProperties.active_fallback = node.id;
-      AppProperties.state = AppProperties.STATES.EDIT_FALLBACK
-      return;
+      AppProperties.state = AppProperties.STATES.EDIT_FALLBACK;
+      break;
     // default:
     //   return AppProperties.state=AppProperties.STATES.HOME;
   }
+  // LFS state takes precedence over other states
+  if (node.isLFSPointer) {
+    AppProperties.active_lfs_file = node.id;
+    AppProperties.state = AppProperties.STATES.EDIT_LFS;
+  }
+  AppProperties.active_node = node;
 };
 
 const updatePath = async ([path,type]) => {
@@ -552,86 +552,9 @@ const openFileWithDefaultApplication = async (node: ArcTreeViewNode) => {
   await window.ipc.invoke('LocalFileSystemService.openFileNative', p);
 };
 
+
 const downloadLFSFiles = async paths => {
-
-  if(!AppProperties.user)
-    return $q.dialog({
-      component: ConfirmationDialog,
-      componentProps: {
-        title: 'Authentication Error',
-        msg: 'You need to be logged in to download LFS files.'
-      }
-    });
-
-  let response = null;
-
-  // get remote name and url
-  response = await window.ipc.invoke('GitService.run', {
-    args: [`remote`],
-    cwd: ArcControlService.props.arc_root
-  });
-  if(!response[0]) return $q.dialog({
-    component: ConfirmationDialog,
-    componentProps: {
-      title: 'Error',
-      msg: 'Unable to determine remote name'
-    }
-  });
-  const remote_name = response[1].split('\n')[0];
-  response = await window.ipc.invoke('GitService.run', {
-    args: [`remote`,`get-url`,remote_name],
-    cwd: ArcControlService.props.arc_root
-  });
-  if(!response[0]) return $q.dialog({
-    component: ConfirmationDialog,
-    componentProps: {
-      title: 'Error',
-      msg: 'Unable to determine remote url'
-    }
-  });
-  const remote_url = response[1].split('\n')[0];
-
-  // patch remote
-  const patched_remote_url = GitService.patch_remote(remote_url);
-
-  response = await window.ipc.invoke('GitService.run', {
-    args: [`remote`,`set-url`,remote_name,patched_remote_url],
-    cwd: ArcControlService.props.arc_root
-  });
-  if(!response[0]) return;
-
-  const dialogProps = reactive({
-    title: 'Pulling Individual LFS Files',
-    ok_title: 'Ok',
-    cancel_title: null,
-    state: 0,
-  });
-
-  $q.dialog({
-    component: GitDialog,
-    componentProps: dialogProps
-  });
-
-  for(let path of paths)
-    await window.ipc.invoke('GitService.run', {
-      args: ['lfs','pull','--include',`"${path}"`],
-      cwd: ArcControlService.props.arc_root
-    });
-    
-  recUpdateNodes(n=>{
-    if(n.isLFS && paths.includes(n.id_rel)){
-      n.downloaded = true;
-      n.isLFSPointer = false;
-    }
-  },props.nodes);
-
-  // unpatch
-  response = await window.ipc.invoke('GitService.run', {
-    args: [`remote`,`set-url`,remote_name,remote_url],
-    cwd: ArcControlService.props.arc_root
-  });
-
-  dialogProps.state=1;
+  await useDownloadLFSFiles($q, paths, props.nodes);
 };
 
 const onCellContextMenu = async (e,node: ArcTreeViewNode) => {
@@ -717,14 +640,17 @@ const onCellContextMenu = async (e,node: ArcTreeViewNode) => {
     items.push({
       label: "Download LFS Files",
       icon: h( 'i', icon_style, ['cloud_download'] ),
-      onClick: ()=>downloadLFSFiles([node.id_rel+'/**'])
+      onClick: () => downloadLFSFiles([node.id_rel ? node.id_rel + '/**' : ''])
     });
   } else {
     if(node.isLFSPointer){
       items.push({
         label: "Download LFS File",
         icon: h( 'i', icon_style, ['cloud_download'] ),
-        onClick: ()=>downloadLFSFiles([node.id_rel])
+        onClick: async () => {
+          await downloadLFSFiles([node.id_rel]);
+          triggerNode({}, node);
+        }
       });
     }
   }
@@ -857,7 +783,7 @@ const onCellContextMenu = async (e,node: ArcTreeViewNode) => {
       icon: h( 'i', icon_style, ['delete'] ),
       onClick: ()=>confirm_delete(node,()=>ArcControlService.delete('GetRunRemoveContracts',node.label))
     });
-  } else if (node.type===formatNodeEditString(Datamap)){ 
+  } else if (node.type===formatNodeEditString(Datamap)){
     items.push({
         label: "Delete",
         icon: h( 'i', icon_style, ['delete'] ),
@@ -996,6 +922,13 @@ watch(()=>AppProperties.state, async (newValue, oldValue) => {
     AppProperties.STATES.VALIDATION
   ].includes(newValue)){
     props.selection = '';
+  }
+});
+
+watch(()=> AppProperties.node_needs_refresh, async (newValue, oldValue) => {
+  if (newValue) {
+    triggerNode({}, AppProperties.active_node)
+    AppProperties.node_needs_refresh = false;
   }
 });