+++ CHANGELOG.md
@@ -25,6 +25,7 @@ where applicable. Add your name in the style of `(by @github-username)` at the e
 - Files without view support by ARCitect will open a default view with a button allowing the user to open them with the default application. This feature was also added for all files in the context menu #455 (by @Thyra, @Freymaurer)
 - Auto save changes via Markdown editor when switching files or closing Markdown editor #428 (by @Thyra)
 - Added error handling for file write operations. This will now show a modal if a file could not be written to disc, instead of printing to console. This can for example happen if a file should be written that is already open in another tool. (by @Freymaurer)
+- When selecting LFS files which are not downloaded yet, ARCitect will display a view explaining that and allowing the user to download the file #466 (by @Thyra)
 
 ### 🔄 Changed